readequando codigo para entender segfault

+++ ft_print_char.c
@@ -6,13 +6,13 @@
 /*   By: apaula-b <apaula-b@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/06/24 11:03:20 by apaula-b          #+#    #+#             */
-/*   Updated: 2021/06/26 14:33:53 by apaula-b         ###   ########.fr       */
+/*   Updated: 2021/06/28 21:52:16 by apaula-b         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-size_t	print_char(va_list args)
+int	print_char(va_list args)
 {
 	char	data_to_print;
